pyproject.toml
[project]
name = "live-coding-langchain-langgraph"
version = "0.1.0"
description = "A live coding project for LangChain and LangGraph."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.10",
    "langchain-openai>=1.1.10",
    "tavily-python>=0.7.21",
    "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
dev = [
    "langgraph-cli[inmem]>=0.4.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/chef_agent"]
